Moved Writer from Support to Fixture
This is more a unit testing terminology.
